feat: ゲーム状態管理のためのカスタムフック useGameState を追加し、ホームページのテストをリファクタリング
